+++ source/radTerminal.Info.pas
@@ -3,7 +3,7 @@ unit radTerminal.Info;
 interface
 
 const
-  Version = '0.5.8.0';
+  Version = '0.6.8.0';
 
 implementation